refactor: pieces 패키지 생성 및 구조 변경
- pawn 클래스 pieces 패키지로 이동
