+++ src/test/java/wootecamp/chess/BoardTest.java
@@ -3,10 +3,12 @@ package wootecamp.chess;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wootecamp.chess.pieces.Pawn;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class BoardTest {
+
     private Board board;
 
     @BeforeEach
