+++ src/test/java/wootecamp/chess/pieces/PawnTest.java
@@ -1,4 +1,4 @@
-package wootecamp.chess;
+package wootecamp.chess.pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
